Server list modificiations
I think I did some changes to how the session list works, but I couldn't find what I was looking for, so I'm done for today. For reference, I wanted to make it so when you click on a session in the list of session, you're able to get more information than just the name and the ping, things like which players are joined (not just numbers, steam usernames) and other info.